check the like,comment

+++ controllers/commentsController.js
@@ -1,6 +1,7 @@
 const Comment = require("../models/commentsModel")
 const AppError = require("../utils/AppError")
 const catchAsync = require("../utils/catchAsync")
+const checkObject = require("../utils/checkFile")
 
 const postNewComment = catchAsync(async(req,res,next)=>{
     const {text} = req.body
@@ -15,6 +16,8 @@ const postNewComment = catchAsync(async(req,res,next)=>{
 })
 //  delete comment
 const deleteComment = catchAsync(async(req,res,next)=>{
+    if(!await checkObject(Comment,req)) return next(new AppError(400,"you have no comment on this file"))
+
     const comment = await Comment.findOneAndDelete({user:req.user._id,file:req.params.fileId})
     res.status(200).send("deleted")
     next()
